// src/generated/openapi/models/BenchmarkStatusDTO.ts
// @ts-nocheck - Auto-generated file, TypeScript strict checks disabled intentionally
/* tslint:disable */
/* eslint-disable */
/**
 * VIPLEV
 * This is an API specification
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BenchmarkStatusDTO
 */
export interface BenchmarkStatusDTO {
    /**
     * ID of the benchmark
     * @type {string}
     * @memberof BenchmarkStatusDTO
     */
    benchmarkId?: string;
    /**
     * ID of the run
     * @type {string}
     * @memberof BenchmarkStatusDTO
     */
    runId?: string;
    /**
     * Current status of the run
     * @type {BenchmarkStatusDTOStatusEnum}
     * @memberof BenchmarkStatusDTO
     */
    status?: BenchmarkStatusDTOStatusEnum;
}


/**
 * @export
 */
export const BenchmarkStatusDTOStatusEnum = {
    PendingStart: 'PENDING_START',
    Started: 'STARTED',
    PendingStop: 'PENDING_STOP',
    Stopped: 'STOPPED',
    Finished: 'FINISHED',
    Failed: 'FAILED'
} as const;
export type BenchmarkStatusDTOStatusEnum = typeof BenchmarkStatusDTOStatusEnum[keyof typeof BenchmarkStatusDTOStatusEnum];


/**
 * Check if a given object implements the BenchmarkStatusDTO interface.
 */
export function instanceOfBenchmarkStatusDTO(value: object): value is BenchmarkStatusDTO {
    return true;
}

export function BenchmarkStatusDTOFromJSON(json: any): BenchmarkStatusDTO {
    return BenchmarkStatusDTOFromJSONTyped(json, false);
}

export function BenchmarkStatusDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): BenchmarkStatusDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'benchmarkId': json['benchmarkId'] == null ? undefined : json['benchmarkId'],
        'runId': json['runId'] == null ? undefined : json['runId'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function BenchmarkStatusDTOToJSON(json: any): BenchmarkStatusDTO {
    return BenchmarkStatusDTOToJSONTyped(json, false);
}

export function BenchmarkStatusDTOToJSONTyped(value?: BenchmarkStatusDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'benchmarkId': value['benchmarkId'],
        'runId': value['runId'],
        'status': value['status'],
    };
}